server/requirements.txt
fastapi[standard]
uvicorn
groq
pydantic
python-dotenv
